DRILL-4785: Avoid checking for hard affinity scans in limit 0 shortcut cases
